pass queries to AssociationCounter
